requirements.txt
mysql-connector==2.2.9
numpy==1.16.4
opencv-python==3.2.0.8
selenium==3.141.0
urllib3==1.25.3
